refactor: preview 분리
